feat(projen.component.ts-ast): extract `utils.projen` to `utils.ts-ast` modules
